fix: change the type of calorie limit from integer to float

+++ app/src/main/java/com/example/ingredientscanner/data/remote/models/NutritionixItemResponse.java
@@ -4,5 +4,5 @@ public class NutritionixItemResponse {
     public String item_name;
     public String brand_name;
     public String nf_ingredient_statement;
-    public Double nf_calories;
+    public Integer nf_calories;
 }
